+++ libsolidity/ast/Types.cpp
@@ -3445,13 +3445,15 @@ FunctionTypePointer FunctionType::asExternallyCallableFunction(bool _inLibrary, 
 
 	TypePointers parameterTypes;
 	for (auto const& t: m_parameterTypes)
-	{
-		auto refType = dynamic_cast<ReferenceType const*>(t);
-		if (refType && refType->location() == DataLocation::CallData)
-			parameterTypes.push_back(TypeProvider::withLocation(refType, DataLocation::Memory, true));
-		else
-			parameterTypes.push_back(t);
-	}
+		parameterTypes.push_back(
+			TypeProvider::withLocationIfReference(DataLocation::CallData, DataLocation::Memory, t, true)
+		);
+
+	TypePointers returnParameterTypes;
+	for (auto const& returnParamType: m_returnParameterTypes)
+		returnParameterTypes.push_back(
+			TypeProvider::withLocationIfReference(DataLocation::CallData, DataLocation::Memory, returnParamType, true)
+		);
 
 	Kind kind = m_kind;
 	if (_inLibrary)
@@ -3465,7 +3467,7 @@ FunctionTypePointer FunctionType::asExternallyCallableFunction(bool _inLibrary, 
 
 	return TypeProvider::function(
 		parameterTypes,
-		m_returnParameterTypes,
+		returnParameterTypes,
 		m_parameterNames,
 		m_returnParameterNames,
 		kind,